+++ src/fr/maxlego08/menu/api/button/PaginateButton.java
@@ -1,0 +1,7 @@
+package fr.maxlego08.menu.api.button;
+
+public interface PaginateButton extends Button {
+
+    int getPaginationSize();
+
+}