+++ changelog.md
@@ -35,6 +35,7 @@
 # Unreleased
 
 - Fix NmsVersion with 1.8 and 1.9
+- Fix anti-dupe for old version
 - Fix MenuItemStack loader with placeholder check
 - Fix CONTAINS_STRING action for placeholder
 - Add matrix support for button slots [#39](https://github.com/Maxlego08/zMenu/issues/39)